pyproject.toml
[project]
name = "vision-os-backend"
version = "0.1.0"
description = "The OS to design your life."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-unfold>=0.90.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pydantic[email]>=2.13.3",
]
